Maybe floatify printed metrics
Rationals are, IMO, sometimes hard to interpret. This PR introduces a
variable `*print-rationals*`, which when `T` does nothing special --
metrics are printed as they always have been; when `nil`,
`print-table` will convert metrics of type `(and rational (not
integer))` to `float`s, which I believe are much easier to read
at-a-glance.

This does not affect the underlying data (e.g. metrics in a `timer`
object) -- only the printed representation (via `report` or
`print-table`).