call `force-output` when printing verbose messages
When running benchmarks in Sly with `*trace-output*` bound to the SLY mrepl,
`*trace-output*` is line-buffered, which means I only see the "~&Benchmarking ~S..."
message after the benchmark has completed, when "done~%" adds a newline. This makes the
before/after messages kinda useless.

This commit inserts a call to `force-output` after printing the before message, but before
running the benchmark.